Stop pred_bm's int16 cast from turning NaN into 0
Unpredictable pixels are deliberately left NaN in the float32 output,
but the final .astype(np.int16) silently rounds NaN down to 0 --
confirmed leaking into real output wherever a derived index (RDVI,
SATVI, etc.) went NaN/Inf even though NIR1 itself was fine, since the
caller's re-mask only checks NIR1's own nullity. Return float32
instead, matching pred_cov/pred_cp's own pattern; nothing downstream
needed int16.

+++ hlsstack/hls_funcs/predict.py
@@ -77,7 +77,7 @@ def pred_bm(dat, model):
             del mat_t, df_t, valid_mask
 
         del mat
-        return out.astype(np.int16).reshape((time_steps,) + grid)
+        return out.reshape((time_steps,) + grid)
 
     def pred_func_xr(dat_xr, model_vars_xr):
         # The index functions run on the native (time, y, x) grid. They used to
